massive refactors!!!

+++ source/modus_core/detail/Duration.hpp
@@ -7,21 +7,21 @@ namespace ml
 {
 	/* * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
 
-	ML_alias nanoseconds_t	= typename chrono::duration<float64_t, std::nano>;
-	ML_alias microseconds_t	= typename chrono::duration<float64_t, std::micro>;
-	ML_alias milliseconds_t	= typename chrono::duration<float64_t, std::milli>;
-	ML_alias seconds_t		= typename chrono::duration<float64_t, std::ratio<1>>;
-	ML_alias minutes_t		= typename chrono::duration<float64_t, std::ratio<60>>;
-	ML_alias hours_t		= typename chrono::duration<float64_t, std::ratio<60 * 60>>;
-	ML_alias days_t			= typename chrono::duration<float64_t, std::ratio<60 * 60 * 24>>;
+	ML_alias nanoseconds_t	= typename chrono::duration<float32, std::nano>;
+	ML_alias microseconds_t	= typename chrono::duration<float32, std::micro>;
+	ML_alias milliseconds_t	= typename chrono::duration<float32, std::milli>;
+	ML_alias seconds_t		= typename chrono::duration<float32, std::ratio<1>>;
+	ML_alias minutes_t		= typename chrono::duration<float32, std::ratio<60>>;
+	ML_alias hours_t		= typename chrono::duration<float32, std::ratio<60 * 60>>;
+	ML_alias days_t			= typename chrono::duration<float32, std::ratio<60 * 60 * 24>>;
 
 	/* * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
 
-	struct duration final : public chrono::duration<float64_t>
+	struct duration final : public chrono::duration<float32>
 	{
 		/* * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
 
-		using chrono::duration<float64_t>::duration;
+		using chrono::duration<float32>::duration;
 
 		/* * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
 
@@ -70,7 +70,7 @@ namespace ml
 
 	static void from_json(json const & j, duration & v)
 	{
-		j.get_to((float64_t &)v);
+		j.get_to((float32 &)v);
 	}
 
 	/* * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
